Refactor project.
Provide fixes for previous task.
