+++ js/modules/adverts/index.js
@@ -1,0 +1,12 @@
+import createSingleAdvert from './create-single-advert.js';
+import createAdverts from './create-adverts.js';
+
+// console.debug('index adverts');
+function testAdverts() {
+  return 'index adverts';
+}
+function testAdverts2() {
+  return 'index adverts 2';
+}
+
+export default { testAdverts, testAdverts2, createAdverts, createSingleAdvert };
