+++ js/modules/index.js
@@ -1,0 +1,9 @@
+// import adverts from './adverts.js';
+import adverts from './adverts/index.js';
+
+function testModules() {
+  return 'index modules';
+}
+
+
+export default { testModules, adverts };
