+++ js/utils/index.js
@@ -1,0 +1,3 @@
+import { randomInteger, randomFloat } from './get-random-number.js';
+
+export default { randomInteger, randomFloat };
